+++ controllers/certificate/certificate_controller.go
@@ -18,8 +18,6 @@ import (
 
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
-	apierrs "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/client-go/util/cert"
 	ctrl "sigs.k8s.io/controller-runtime"
@@ -40,77 +38,6 @@ const (
 
 	apiRuleCRDName = "apirules.gateway.kyma-project.io"
 )
-
-var currentCertificate *tls.Certificate
-
-func GetCertificate(*tls.ClientHelloInfo) (*tls.Certificate, error) {
-	if currentCertifate == nil {
-		return nil, errors.New("certificate not available")
-	}
-	return currentCertifate, nil
-}
-
-func InitialiseCertificateSecret(ctx context.Context, client client.Client, log logr.Logger) error {
-	log.Info("Initialising certificate secret", "namespace", secretNamespace, "name", secretName)
-
-	secret := &corev1.Secret{}
-	err := client.Get(ctx, types.NamespacedName{Namespace: secretNamespace, Name: secretName}, secret)
-	if err != nil {
-		if apierrs.IsNotFound(err) {
-			log.Info("Certificate secret not found, creating a new one")
-			certificate, key, err := generateNewCertificate(serviceName, secretNamespace)
-			if err != nil {
-				return errors.Wrap(err, "failed to generate certificate")
-			}
-			secret = &corev1.Secret{
-				ObjectMeta: metav1.ObjectMeta{Namespace: secretNamespace, Name: secretName},
-				Data: map[string][]byte{
-					certificateName: certificate,
-					keyName:         key,
-				},
-				Type: corev1.SecretTypeOpaque,
-			}
-			if err := client.Create(ctx, secret); err != nil {
-				return errors.Wrap(err, "failed to create secret")
-			}
-			if err := updateCertificateInCRD(ctx, client, certificate); err != nil {
-				return errors.Wrap(err, "failed to update certificate into CRD")
-			}
-		} else {
-			return errors.Wrap(err, "failed to get certificate secret")
-		}
-	} else {
-		log.Info("Certificate secret found", "namespace", secretNamespace, "name", secretName)
-	}
-
-	if err = parseCertificateSecret(secret, log); err != nil {
-		return errors.Wrap(err, "failed to get parse certificate secret")
-	}
-
-	return nil
-}
-
-func ReadCertificateSecret(ctx context.Context, client client.Client, log logr.Logger) error {
-	log.Info("Reading certficate secret", "namespace", secretNamespace, "name", secretName)
-
-	secret := &corev1.Secret{}
-	err := client.Get(ctx, types.NamespacedName{Namespace: secretNamespace, Name: secretName}, secret)
-	if err != nil {
-		return errors.Wrap(err, "failed to get certificate secret")
-	}
-
-	if err = parseCertificateSecret(secret, log); err != nil {
-		return errors.Wrap(err, "failed to get parse certificate secret")
-	}
-
-	tlsCert, err := tls.X509KeyPair(secret.Data[certificateName], secret.Data[keyName])
-	if err != nil {
-		return errors.Wrap(err, "failed to load certificate key pair")
-	}
-	currentCertifate = &tlsCert
-
-	return nil
-}
 
 func NewCertificateReconciler(mgr manager.Manager) *Reconciler {
 	return &Reconciler{
@@ -141,7 +68,7 @@ func (r *Reconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Resu
 	if err != nil {
 		return ctrl.Result{Requeue: false}, err
 	}
-	currentCertifate = &tlsCert
+	currentCertificate = &tlsCert
 
 	return ctrl.Result{RequeueAfter: reconciliationInterval}, nil
 }
